runc update: fix update of rdt schemata
Fix an issue with --l3-cache-schema and --mem-bw-schema where the
resctrlfs schemata was not updated if the original spec didn't have the
intelRdt field specified. Confusingly, calling runc update twice did the
job: the first call created the resctrl group and the seccond call was
able to update the schemata.

NOTE: A serious flaw still remains in the runc update command with this
functionality. If new resctrl group is created (i.e. the original
intelRdt config was empty), runc only puts the init process to the
new resctrl group, all child processes being left to the original group,
escaping the update. An alternative approach could be to make the update
fail in this scenario until completely fixed in future.