libcontainer/intelrdt: create clos on Set()
Create the clos (resctrl group) in the Set() method which configures the
schemata for the group. Removes the interdependency between Apply() and
Set() where Apply() (which assigns a pid to the clos) was required to be
called before Set() to make sure that the clos existed before writing
it's schemata.